+++ src/xSTIR/cSTIR/include/sirf/STIR/cstir.h
@@ -99,6 +99,9 @@ extern "C" {
         (const void* ptr_acq_matrix,
         const float collimator_sigma_0_in_mm, const float collimator_slope_in_mm, const bool full_3D);
 
+    // listmode data methods
+    void* cSTIR_acquisitionDataFromListmode(void* ptr_t);
+
 	// Acquisition data methods
 	void* cSTIR_getAcquisitionDataStorageScheme();
 	void* cSTIR_setAcquisitionDataStorageScheme(const char* scheme);
@@ -121,7 +124,8 @@ extern "C" {
 	void* cSTIR_fillAcquisitionDataFromAcquisitionData
 		(void* ptr_acq, const void * ptr_from);
 	void* cSTIR_writeAcquisitionData(void* ptr_acq, const char* filename);
-	void* cSTIR_get_ProjDataInfo(void* ptr_acq);
+        // works for AcquisitionData, ListmodeData and ImageData at present
+	void* cSTIR_get_info(void* ptr_acq);
 	void* cSTIR_get_subset(void* ptr_acq, int nv, size_t ptr_views);
 
 	// Reconstruction methods